+++ application/models/FleetModel.php
@@ -4,7 +4,8 @@ defined('BASEPATH') OR exit('No direct script access allowed');
 
 class FleetModel extends CSV_Model
 {
-    private $budget = 10000000;
+    private $budget = 10000;
+    private $plane_types = null;
 
     // Constructor
     public function __construct()
@@ -109,6 +110,30 @@ class FleetModel extends CSV_Model
             $grandTotal += doubleval($planes[$value->model_id]['price']);
         }
         return $grandTotal;
+    }
+    
+    /**
+     * Check if the given plane type is valid
+     * @param type $type
+     */
+    public function checkPlaneType($type){
+        if($this->plane_types == null){
+            $wackyModel = new WackyModel();
+            $planes = $wackyModel->airplanes();
+            $this->plane_types = array();
+            foreach ($planes as $key => $plane) {
+                $this->plane_types[$plane['id']] = $plane['id'];
+            }
+            return in_array($type, $this->plane_types);
+        }
+    }
+    
+    /**
+     * Check if the budget is sufficient to make a purchase of certain price
+     * @param type $price price
+     */
+    public function checkSufficientBudget($price){
+        return $this->getBudget() - $this->getGrandTotal() - $price >= 0;
     }
     
     /**