+++ CHANGELOG.md
@@ -13,6 +13,12 @@ All notable changes to this project will be documented in this file.
 
 # [latest stable version 2.0.0]
 
+## [3.0.0] - 2017-11-23
+## Changed
+ - Add unit test against business logic on fleet.
+ - Update unit test on flight due to the change of data structure.
+ - Add checking for sufficient budget on purchase.
+
 ## [3.0.0] - 2017-11-22
 ## Changed
  - Display more information on fleet page and plane page.